refactor(): rename post module :bug:
