+++ src/routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from 'nest-router';
-import { PostsModule } from './v1/controllers/posts/posts.module';
+import { PostsModule } from './v1/controllers/posts/post.module';
 import { UserModule } from './v1/controllers/users/user.module';
 import { V1Module } from './v1/v1.module';
