+++ src/v1/v1.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UserModule } from './controllers/users/user.module';
-import { PostsModule } from './controllers/posts/posts.module';
+import { PostsModule } from './controllers/posts/post.module';
 
 @Module({
   imports: [UserModule, PostsModule],
